move kernel stack management to rust

+++ kernel/trap.c
@@ -115,8 +115,8 @@ void usertrapret(void) {
 
   // set up trapframe values that uservec will need when
   // the process next traps into the kernel.
-  p->trapframe->kernel_satp   = r_satp();  // kernel page table
-  p->trapframe->kernel_sp     = p->kstack; // process's kernel stack
+  p->trapframe->kernel_satp   = r_satp();       // kernel page table
+  p->trapframe->kernel_sp     = (u64)p->kstack; // process's kernel stack
   p->trapframe->kernel_trap   = (u64)usertrap;
   p->trapframe->kernel_hartid = r_tp(); // hartid for cpuid()
